[openssl] Drop make and perl dependencies from openssl package. JB#63112
The makefile seemed there for SOURCE2 which we don't install and
regarding that Fedora moved it around a bit and removed the make
dependency with f20f5f466f6a in 2017.

c_rehash is now duplicate to the one in -perl subpackage and appears
introduced by update to 1.1.1g by commit f8a4976b.